Honestly I don't know
I don’t think i’ve made any progress this time around. Except that I
know how the level will end. Lasers and hazards are far too involved.
Timing much too precise for a second level. It’s going to be a simple
team cube toss ftw. Haven’t decided if it’ll just hit a button, or if
it’ll block a laser and allow the final door to open whilst blocking.

but i still need to do it